+++ packages/gatsby-plugin-image/src/components/static-image.tsx
@@ -1,10 +1,12 @@
-import { FluidObject, FixedObject } from "gatsby-image"
-import { GatsbyImage as GatsbyImageBrowser } from "./gatsby-image.browser"
+import {
+  GatsbyImage as GatsbyImageBrowser,
+  ISharpGatsbyImageData,
+} from "./gatsby-image.browser"
 import { _getStaticImage } from "./static-image.server"
 import { StaticImageProps } from "../utils"
 // These values are added by Babel. Do not add them manually
 interface IPrivateProps {
-  __imageData?: FluidObject & FixedObject
+  __imageData?: ISharpGatsbyImageData
   __error?: string
 }