+++ packages/gatsby-plugin-sass/src/__tests__/gatsby-node.js
@@ -58,9 +58,10 @@ describe(`gatsby-plugin-sass`, () => {
 })
 
 describe(`pluginOptionsSchema`, () => {
-  it(`should provide meaningful errors when fields are invalid`, () => {
+  it(`should provide meaningful errors when fields are invalid`, async () => {
     const expectedErrors = [
       `"implementation" must be of type object`,
+      `"cssLoaderOptions" must be of type object`,
       `"postCssPlugins" must be an array`,
       `"sassRuleTest" must be of type object`,
       `"sassRuleModulesTest" must be of type object`,
@@ -85,9 +86,10 @@ describe(`pluginOptionsSchema`, () => {
       `"sourceMapRoot" must be a string`,
     ]
 
-    const { errors } = testPluginOptionsSchema(pluginOptionsSchema, {
+    const { errors } = await testPluginOptionsSchema(pluginOptionsSchema, {
       implementation: `This should be a require() thing`,
       postCssPlugins: `This should be an array of postCss plugins`,
+      cssLoaderOptions: `This should be an object of css-loader options`,
       sassRuleTest: `This should be a regexp`,
       sassRuleModulesTest: `This should be a regexp`,
       useResolveUrlLoader: `This should be a boolean`,
@@ -114,9 +116,10 @@ describe(`pluginOptionsSchema`, () => {
     expect(errors).toEqual(expectedErrors)
   })
 
-  it(`should validate the schema`, () => {
-    const { isValid } = testPluginOptionsSchema(pluginOptionsSchema, {
+  it(`should validate the schema`, async () => {
+    const { isValid } = await testPluginOptionsSchema(pluginOptionsSchema, {
       implementation: require(`../gatsby-node.js`),
+      cssLoaderOptions: { camelCase: false },
       postCssPlugins: [{ post: `CSS plugin` }],
       sassRuleTest: /\.global\.s(a|c)ss$/,
       sassRuleModulesTest: /\.mod\.s(a|c)ss$/,