+++ packages/gatsby-plugin-sharp/src/index.js
@@ -1,5 +1,5 @@
 const sharp = require(`./safe-sharp`)
-
+const { generateImageData } = require(`./image-data`)
 const imageSize = require(`probe-image-size`)
 
 const _ = require(`lodash`)
@@ -18,8 +18,32 @@ const {
 const { memoizedTraceSVG, notMemoizedtraceSVG } = require(`./trace-svg`)
 const duotone = require(`./duotone`)
 const { IMAGE_PROCESSING_JOB_NAME } = require(`./gatsby-worker`)
+const { getDimensionsAndAspectRatio } = require(`./utils`)
+// const { rgbToHex } = require(`./utils`)
 
 const imageSizeCache = new Map()
+
+const getImageSizeAsync = async file => {
+  if (
+    process.env.NODE_ENV !== `test` &&
+    imageSizeCache.has(file.internal.contentDigest)
+  ) {
+    return imageSizeCache.get(file.internal.contentDigest)
+  }
+  const input = fs.createReadStream(file.absolutePath)
+  const dimensions = await imageSize(input)
+
+  if (!dimensions) {
+    reportError(
+      `gatsby-plugin-sharp couldn't determine dimensions for file:\n${file.absolutePath}\nThis file is unusable and is most likely corrupt.`,
+      ``
+    )
+  }
+
+  imageSizeCache.set(file.internal.contentDigest, dimensions)
+  return dimensions
+}
+// Remove in next major as it's really slow
 const getImageSize = file => {
   if (
     process.env.NODE_ENV !== `test` &&
@@ -27,9 +51,7 @@ const getImageSize = file => {
   ) {
     return imageSizeCache.get(file.internal.contentDigest)
   } else {
-    const dimensions = imageSize.sync(
-      toArray(fs.readFileSync(file.absolutePath))
-    )
+    const dimensions = imageSize.sync(fs.readFileSync(file.absolutePath))
 
     if (!dimensions) {
       reportError(
@@ -53,65 +75,11 @@ exports.setBoundActionCreators = actions => {
   boundActionCreators = actions
 }
 
+exports.generateImageData = generateImageData
+
 function calculateImageDimensionsAndAspectRatio(file, options) {
-  // Calculate the eventual width/height of the image.
   const dimensions = getImageSize(file)
-  const imageAspectRatio = dimensions.width / dimensions.height
-
-  let width = options.width
-  let height = options.height
-
-  switch (options.fit) {
-    case sharp.fit.fill: {
-      width = options.width ? options.width : dimensions.width
-      height = options.height ? options.height : dimensions.height
-      break
-    }
-    case sharp.fit.inside: {
-      const widthOption = options.width
-        ? options.width
-        : Number.MAX_SAFE_INTEGER
-      const heightOption = options.height
-        ? options.height
-        : Number.MAX_SAFE_INTEGER
-
-      width = Math.min(widthOption, Math.round(heightOption * imageAspectRatio))
-      height = Math.min(
-        heightOption,
-        Math.round(widthOption / imageAspectRatio)
-      )
-      break
-    }
-    case sharp.fit.outside: {
-      const widthOption = options.width ? options.width : 0
-      const heightOption = options.height ? options.height : 0
-
-      width = Math.max(widthOption, Math.round(heightOption * imageAspectRatio))
-      height = Math.max(
-        heightOption,
-        Math.round(widthOption / imageAspectRatio)
-      )
-      break
-    }
-
-    default: {
-      if (options.width && !options.height) {
-        width = options.width
-        height = Math.round(options.width / imageAspectRatio)
-      }
-
-      if (options.height && !options.width) {
-        width = Math.round(options.height * imageAspectRatio)
-        height = options.height
-      }
-    }
-  }
-
-  return {
-    width,
-    height,
-    aspectRatio: width / height,
-  }
+  return getDimensionsAndAspectRatio(dimensions, options)
 }
 
 function prepareQueue({ file, args }) {
@@ -432,7 +400,6 @@ async function stats({ file, reporter }) {
 
 async function fluid({ file, args = {}, reporter, cache }) {
   const options = healOptions(getPluginOptions(), args, file.extension)
-
   if (options.sizeByPixelDensity) {
     /*
      * We learned that `sizeByPixelDensity` is only valid for vector images,
@@ -657,7 +624,7 @@ async function fixed({ file, args = {}, reporter, cache }) {
   sizes.push(options[fixedDimension])
   sizes.push(options[fixedDimension] * 1.5)
   sizes.push(options[fixedDimension] * 2)
-  const dimensions = getImageSize(file)
+  const dimensions = await getImageSizeAsync(file)
 
   const filteredSizes = sizes.filter(size => size <= dimensions[fixedDimension])
 
@@ -763,17 +730,8 @@ async function fixed({ file, args = {}, reporter, cache }) {
   }
 }
 
-function toArray(buf) {
-  var arr = new Array(buf.length)
-
-  for (var i = 0; i < buf.length; i++) {
-    arr[i] = buf[i]
-  }
-
-  return arr
-}
-
 exports.queueImageResizing = queueImageResizing
+exports.batchQueueImageResizing = batchQueueImageResizing
 exports.resize = queueImageResizing
 exports.base64 = base64
 exports.generateBase64 = generateBase64
@@ -783,4 +741,5 @@ exports.resolutions = fixed
 exports.fluid = fluid
 exports.fixed = fixed
 exports.getImageSize = getImageSize
+exports.getImageSizeAsync = getImageSizeAsync
 exports.stats = stats