+++ packages/gatsby-plugin-feed/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.20...gatsby-plugin-feed@2.6.0) (2020-11-02)
+
+### Features
+
+- **gatsby:** release plugin option validation ([#27437](https://github.com/gatsbyjs/gatsby/issues/27437)) ([41ae1c0](https://github.com/gatsbyjs/gatsby/commit/41ae1c07ad9919655782ef17feed8cf4f14f12d8))
+
+## [2.5.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.19...gatsby-plugin-feed@2.5.20) (2020-10-28)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
+## [2.5.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.18...gatsby-plugin-feed@2.5.19) (2020-10-22)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
+## [2.5.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.17...gatsby-plugin-feed@2.5.18) (2020-10-21)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
+## [2.5.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.16...gatsby-plugin-feed@2.5.17) (2020-10-20)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
+## [2.5.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.15...gatsby-plugin-feed@2.5.16) (2020-10-19)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
 ## [2.5.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.5.14...gatsby-plugin-feed@2.5.15) (2020-10-16)
 
 **Note:** Version bump only for package gatsby-plugin-feed