+++ packages/gatsby-plugin-offline/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.3.0...gatsby-plugin-offline@3.3.1) (2020-11-03)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+# [3.3.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.38...gatsby-plugin-offline@3.3.0) (2020-11-02)
+
+### Features
+
+- **gatsby:** release plugin option validation ([#27437](https://github.com/gatsbyjs/gatsby/issues/27437)) ([41ae1c0](https://github.com/gatsbyjs/gatsby/commit/41ae1c07ad9919655782ef17feed8cf4f14f12d8))
+
+## [3.2.38](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.37...gatsby-plugin-offline@3.2.38) (2020-10-28)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+## [3.2.37](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.36...gatsby-plugin-offline@3.2.37) (2020-10-22)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+## [3.2.36](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.35...gatsby-plugin-offline@3.2.36) (2020-10-21)
+
+### Bug Fixes
+
+- Update plugin schema testing util and associated tests ([#27574](https://github.com/gatsbyjs/gatsby/issues/27574)) ([6d81283](https://github.com/gatsbyjs/gatsby/commit/6d81283e4f47ae2cb571626bf4d02fcd2c9d1af4))
+
+## [3.2.35](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.34...gatsby-plugin-offline@3.2.35) (2020-10-20)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
+## [3.2.34](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.33...gatsby-plugin-offline@3.2.34) (2020-10-19)
+
+**Note:** Version bump only for package gatsby-plugin-offline
+
 ## [3.2.33](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-offline@3.2.32...gatsby-plugin-offline@3.2.33) (2020-10-16)
 
 **Note:** Version bump only for package gatsby-plugin-offline