+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.3.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.39...gatsby-source-shopify@3.3.0) (2020-11-02)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.2.39](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.38...gatsby-source-shopify@3.2.39) (2020-10-27)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.2.38](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.37...gatsby-source-shopify@3.2.38) (2020-10-26)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
 ## [3.2.37](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.2.36...gatsby-source-shopify@3.2.37) (2020-10-16)
 
 **Note:** Version bump only for package gatsby-source-shopify